+++ src/client/services/is-admin.js
@@ -1,0 +1,6 @@
+// Authority 
+import authHeader from '@/services/auth-header.js';
+
+export default function isAdmin() {
+	return !(JSON.stringify(authHeader()) === '{}');
+}