+++ src/client/views/_pages/Dashboard/index.js
@@ -6,7 +6,8 @@ import {
 } from 'react-router-dom';
 import { __ } from '@/components/_utils/_all';
 import Footer from '@/components/Footer/index.tsx';
-import Authorized from '@/pages/Dashboard/Authorized.js';
+import DashBoardComponents from '@/pages/Dashboard/com.index.js';
+
 import { PrivateRoute } from '@/router/PrivateRoute.js';
 
 
@@ -51,7 +52,7 @@ function HookContent() {
 			<hr />
 		
 			  <Switch>
-		        <PrivateRoute exact path="/dashboard" component={Authorized} />
+                <PrivateRoute exact path="/dashboard" component={DashBoardComponents} />
 			  </Switch>
 		</>
 	)